# Implement size_hint for some iterators