--- a/src/cargo/core/workspace.rs
+++ b/src/cargo/core/workspace.rs
@@ -776,6 +776,11 @@ impl<'a, 'cfg> Iterator for Members<'a, 'cfg> {
             }
         }
     }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        let (_, upper) = self.iter.size_hint();
+        (0, upper)
+    }
 }
 
 impl MaybePackage {