This PR implements size_hints for `Deps`, `DepsNotReplaced`, and `Members`.
These size_hints are used extensively by cargo to allocate Vecs.

`Deps`, `DepsNotReplaced`, and `RcVecIter` also now implement `ExactSizeIterator`.

Closes #5211