@Mark-Simulacrum 2018-04-01
You'll want to run `cargo +stable fmt`:

```
Diff in /home/travis/build/rust-lang/cargo/src/cargo/core/resolver/resolve.rs at line 209:
             // Note: Edges is actually a std::collections::hash_set::Iter, which⏎
             // is an ExactSizeIterator.⏎
             Some(ref iter) => iter.size_hint(),⏎
-            None => (0, Some(0))⏎
+            None => (0, Some(0)),⏎
         }⏎
     }⏎
 }⏎
Diff in /home/travis/build/rust-lang/cargo/src/cargo/core/resolver/resolve.rs at line 232:
             // Note: Edges is actually a std::collections::hash_set::Iter, which⏎
             // is an ExactSizeIterator.⏎
             Some(ref iter) => iter.size_hint(),⏎
-            None => (0, Some(0))⏎
+            None => (0, Some(0)),⏎
         }⏎
     }⏎
 }⏎
```