@Phlosioneer 2018-04-01
`cargo +stable fmt` doesn't work (`toolchain ... does not have the bindary 'cargo-fmt'`); and `cargo fmt` changes 80 files. I don't understand what you're trying to do; but I manually did `rustfmt` on just the files I touched for now.