@matklad 2018-04-02
>cargo +stable fmt doesn't work

You probably need `rustup +stable component add rustfmt-preview` for that to work. 

@Phlosioneer did you managed to benchmark this?