@Phlosioneer 2018-04-04
I tried running `cargo bench` but it there aren't any benchmarks to run. How should I benchmark it?