@Phlosioneer 2018-04-05
Alright, I did the benchmarking and some statistics. There's a TL;DR at the bottom.

-----

## Hypothesis

This PR improves `cargo build` speed.

## Method

1) `git checkout master && cargo build --release`. Use `target/release/cargo` for the following steps.
2) clone the target repo.
3) `cargo check`. This forces cargo to download all the dependencies.
4) `cargo clean`. Remove any incremental compilation artifacts from previous runs.
5) `cargo build`.
6) Record the time cargo gives.
7) GOTO 3 until we have 10 data points.
8) `git checkout 5211-missing-size-hints && cargo build --release`. GOTO 2. Use `target/release/cargo`.

## Data

Tests were done on the `serde` and `clippy`. The data-points are listed in order. Some data points were excluded, because the file-system was clearly doing some optimization or something. Points were only excluded if including them would skew *toward* the hypothesis being true.

I tried to also test with `servo`, but it took way too long to get 20 data points. This already took a few hours.

I took all reasonable measures to minimize the load from other programs. The test program was given a higher process priority, and I dedicated exactly two of my computer's four cores to the process.

All numbers are in seconds.

### Serde

```
master branch
27.59 // Thrown out
27.91 // Thrown out
25.80 // Thrown out
23.38
23.42
23.44
23.43
23.14
23.58
23.50
23.45
24.32

5211-missing-size-hints
23.18
23.27
22.92
23.34
23.48
23.27
23.29
23.11
23.17
23.11
```
Calculated values:
```
                    5211-missing-size-hints     master
mean                23.214                      23.378
std. dev.           0.153                       0.147
std. dev. of mean   0.048                       0.047
90% conf. interval  +/- 0.079                   +/- 0.077
max/min             23.23                       23.30
```

### Clippy

```
5211-missing-size-hints
105.83
107.29
107.77
108.10
106.16
106.38
111.83
106.35
118.29
118.75

master
116.75
118.22
118.10
116.16
113.45
111.26
118.23
119.19
108.36
112.00
```
Calculated values:
```
                    5211-missing-size-hints     master
mean                109.68                      115.17
std. dev.           4.967                       3.675
std. dev. of mean   1.571                       1.162
90% conf. interval  +/- 2.584                   +/- 1.911
max/min             112.26                      113.26
```
---
## Conclusion / TL;DR
- For the `serde` test, the maximum bound for this PR was 23.23, while the minimum bound for `master` was 23.30. They do not overlap; therefore, there is a 90% chance (at least) that this speedup was due to this PR.
- For the `clippy` test, the maximum bound for this PR was 112.26, while the minimum bound for `master` was 113.26. They do not overlap; therefore, there is a 90% chance (at least) that this speedup was due to this PR.
- Both the Serde and Clippy tests confirmed the hypothesis independently, with at least a 10% chance (each) of being spurious.
- According to the ranges, performance gain should be at least 1% and at most 10%, on average. Serde saw an average 1% gain, while Clippy saw an average 5% gain.
- `cargo` is remarkably consistent with its compile times.
- `size_hint` is important!
- I have way too much time on my hands.