@Phlosioneer 2018-04-05
P.S. I have a script that did all these calculations for me. I'll clean it up and put it in a PR, for future A/B testing.